Remove components and request scope modules
