--- common/app/com/commercetools/sunrise/common/contexts/RequestScopeModule.java
@@ -1,11 +1,0 @@
-package com.commercetools.sunrise.common.contexts;
-
-import com.google.inject.AbstractModule;
-
-public class RequestScopeModule extends AbstractModule {
-
-    @Override
-    protected void configure() {
-        bindScope(RequestScoped.class, new RequestScope());
-    }
-}
