+++ conf/application.conf
@@ -317,8 +317,6 @@ play.crypto.secret=${?APPLICATION_SECRET}
 play.application.loader = "com.commercetools.sunrise.play.inject.SunriseGuiceApplicationLoader"
 play.http.filters = "com.commercetools.sunrise.play.http.SunriseHttpFilters"
 
-play.modules.enabled += "com.commercetools.sunrise.common.contexts.RequestScopeModule"
-play.modules.enabled += "demo.common.ComponentsModule"
 play.modules.enabled += "com.commercetools.sunrise.common.contexts.ContextDataModule"
 play.modules.enabled += "com.commercetools.sunrise.common.basicauth.BasicAuthModule"
 play.modules.enabled += "com.commercetools.sunrise.common.ctp.CtpModule"
